+++ include/nudb/error.hpp
@@ -81,12 +81,6 @@ namespace errc = boost::system::errc;
 /// Database error codes.
 enum class error
 {
-    /** No error.
-
-        The operation completed successfully.
-    */
-    success = 0,
-
     /** The specified key was not found.
 
         Returned when @ref basic_store::fetch does not